avr-gcc@14: Upgrade to gcc 14.2.0
This updates the GCC sources to 14.2.0 and the corresponding patches by the Darwin GCC maintainer.

The patches by the Darwin GCC maintainer are now applied on non-ARM platforms as well, since the upstream formula for GCC 14 does not make this distinction either.

The `ATMega168pbSupport` option is dropped, as the relevant patches were obsoleted by the recent update of AVR-LibC and were therefore already removed.

Signed-off-by: David Stritzl <437018+davidstritzl@users.noreply.github.com>